move segment cache entries to top level segment-cache dir
