+++ packages/next/src/client/components/segment-cache/cache-key.ts
