+++ packages/next/src/client/components/segment-cache/scheduler.ts
@@ -33,9 +33,9 @@ import { createCacheKey } from './cache-key'
 import {
   FetchStrategy,
   type PrefetchTaskFetchStrategy,
-  getCurrentCacheVersion,
   PrefetchPriority,
-} from '../segment-cache'
+} from './types'
+import { getCurrentCacheVersion } from './cache'
 import {
   addSearchParamsIfPageSegment,
   PAGE_SEGMENT_KEY,
