+++ packages/next/src/client/flight-data-helpers.ts
@@ -9,7 +9,7 @@ import type {
   InitialRSCPayload,
 } from '../shared/lib/app-router-types'
 import { PAGE_SEGMENT_KEY } from '../shared/lib/segment'
-import type { NormalizedSearch } from './components/segment-cache'
+import type { NormalizedSearch } from './components/segment-cache/cache-key'
 import {
   getCacheKeyForDynamicParam,
   parseDynamicParamFromURLPart,
